+++ src/Symfony/Component/Lock/Tests/Store/CombinedStoreTest.php
@@ -32,7 +32,7 @@ class CombinedStoreTest extends AbstractStoreTestCase
     use ExpiringStoreTestTrait;
     use SharedLockStoreTestTrait;
 
-    protected function getClockDelay()
+    protected function getClockDelay(): int
     {
         return 250000;
     }