+++ src/Symfony/Component/Lock/Tests/Store/DoctrineDbalStoreTest.php
@@ -49,7 +49,7 @@ class DoctrineDbalStoreTest extends AbstractStoreTestCase
         @unlink(self::$dbFile);
     }
 
-    protected function getClockDelay()
+    protected function getClockDelay(): int
     {
         return 1000000;
     }